+++ src/ChoyUtilities/Assets/ChoyUtilities/Common/Runtime/Utils/AsyncCollection.cs
@@ -7,7 +7,9 @@ using UnityEngine;
 using UnityEngine.UI;
 
 namespace ChoyUtilities {
+    
     public static partial class HelperCollection {
+        
         private const string TASK_CANCELLATION_MESSAGE = "Task was cancelled";
 
         public static async Awaitable AwaitableUntil(this CancellationToken cancellationToken, Func<bool> condition) {
@@ -24,36 +26,12 @@ namespace ChoyUtilities {
         }
 
         public static async Awaitable RotateObjectAsync(this CancellationToken token,
-            GameObject obj,
-            Vector3 rotateTo,
-            float rotateDuration,
-            Action onDone = null) {
-            try {
-                if (obj is null) return;
-
-                var time = 0f;
-                var startRot = obj.transform.rotation;
-                var endRot = Quaternion.Euler(rotateTo);
-
-                while (time <= rotateDuration) {
-                    time += Time.unscaledDeltaTime;
-                    obj.transform.rotation = Quaternion.Slerp(startRot, endRot, time / rotateDuration);
-                    await Awaitable.EndOfFrameAsync(token);
-                }
-
-                obj.transform.rotation = endRot;
-                onDone?.Invoke();
-            }
-            catch {
-                throw new Exception(TASK_CANCELLATION_MESSAGE);
-            }
+            GameObject obj, float3 rotateTo, float duration, EMotion motion = EMotion.Linear) {
+            await token.RotateObjectAsync(obj.transform, rotateTo, duration, motion);
         }
 
         public static async Awaitable<bool> RotateObjectAsync(this CancellationToken token,
-            Transform obj,
-            float3 rotateTo,
-            float rotateDuration,
-            Action onDone = null) {
+            Transform obj, float3 rotateTo, float duration, EMotion motion = EMotion.Linear) {
             try {
                 if (obj is null) return false;
 
@@ -61,16 +39,15 @@ namespace ChoyUtilities {
                 quaternion startRot = obj.transform.rotation;
                 var endRot = quaternion.Euler(rotateTo);
 
-                while (time <= rotateDuration) {
+                while (time < duration) {
                     time += Time.unscaledDeltaTime;
-                    obj.transform.rotation = math.slerp(startRot, endRot, time / rotateDuration);
+                    obj.transform.rotation = math.slerp(startRot, endRot, 
+                        motion.Evaluate(time / duration));
                     await Awaitable.EndOfFrameAsync(token);
                 }
 
                 if (token.IsCancellationRequested) return false;
                 obj.transform.rotation = endRot;
-                onDone?.Invoke();
-
                 return true;
             }
             catch {
@@ -104,47 +81,15 @@ namespace ChoyUtilities {
         }
 
         public static async Awaitable DialogueAsync(this CancellationToken token,
-            List<string> dialogueList,
-            float dialogueDuration,
-            Action<string> displayTo,
-            float timePerChar,
-            Action onDone = null) {
-            try {
-                if (dialogueList is null || dialogueList.Count == 0) return;
-
-                foreach (var line in dialogueList) {
-                    if (line == string.Empty) continue;
-
-                    var timer = 0f;
-                    var currentDisplaying = "";
-
-                    while (currentDisplaying != line) {
-                        timer += Time.unscaledDeltaTime;
-                        var length = (int)math.ceil(timer / timePerChar);
-                        length = math.clamp(length, 0, line.Length);
-                        currentDisplaying = line[..length];
-                        displayTo(currentDisplaying);
-
-                        await Awaitable.NextFrameAsync(token);
-                    }
-
-                    await Awaitable.WaitForSecondsAsync(dialogueDuration, token);
-                }
-
-                displayTo("");
-                onDone?.Invoke();
-            }
-            catch {
-                throw new Exception(TASK_CANCELLATION_MESSAGE);
-            }
+            List<string> dialogueList, float dialogueDuration,
+            Action<string> displayTo, float timePerChar, Action onDone = null) {
+            await token.DialogueAsync(dialogueList.ToArray(), 
+                dialogueDuration, displayTo, timePerChar, onDone);
         }
 
         public static async Awaitable DialogueAsync(this CancellationToken token,
-            string[] dialogueList,
-            float dialogueDuration,
-            Action<string> displayTo,
-            float timePerChar = 0.05f,
-            Action onDone = null) {
+            string[] dialogueList, float dialogueDuration, Action<string> displayTo,
+            float timePerChar = 0.05f, Action onDone = null) {
             try {
                 if (dialogueList is null || dialogueList.Length == 0) return;
 
@@ -176,13 +121,10 @@ namespace ChoyUtilities {
         }
 
         public static async Awaitable RollRightAngleAsync(this CancellationToken token,
-            Transform ob,
-            float rollSpeed,
-            Vector3 dir,
-            float rollCooldown = 0.1f) {
+            Transform ob, float rollSpeed, float3 dir, float rollCooldown = 0.1f) {
             try {
-                var anchor = ob.position + (Vector3.down + dir) * 0.5f;
-                var axis = Vector3.Cross(Vector3.up, dir);
+                var anchor = (float3)ob.position + (new float3(0, -1f, 0) + dir) * 0.5f;
+                var axis = math.cross(new float3(0, 1f, 0), dir);
 
                 for (var i = 0; i <= 90 / rollSpeed; i++) {
                     ob.RotateAround(anchor, axis, i);
@@ -195,8 +137,7 @@ namespace ChoyUtilities {
         }
 
         public static async Awaitable TimeScaleAsync(this CancellationToken token,
-            float targetScale,
-            float loadDuration = 2f,
+            float targetScale, float loadDuration = 2f, EMotion motion = EMotion.Linear,
             Action onDone = null) {
             try {
                 var currentScale = Time.timeScale;
@@ -204,11 +145,10 @@ namespace ChoyUtilities {
 
                 while (unscaledTimer <= loadDuration) {
                     unscaledTimer += Time.unscaledDeltaTime;
-                    var t = math.lerp(currentScale, targetScale, unscaledTimer / loadDuration);
+                    var t = math.lerp(currentScale, targetScale, motion.Evaluate(unscaledTimer / loadDuration));
                     Time.timeScale = t;
                     await Awaitable.EndOfFrameAsync(token);
                 }
-
                 onDone?.Invoke();
             }
             catch {
@@ -224,58 +164,35 @@ namespace ChoyUtilities {
         }
 
         public static async Awaitable FadeScreenAsync(this CancellationToken token,
-            Image fadeImage,
-            bool isFadein,
-            float loadDuration,
+            Image image, bool isFadein, float duration, EMotion motion = EMotion.Linear,
             Action onDone = null) {
-            try {
-                if (fadeImage is null) return;
-
-                var targetAlpha = isFadein ? 1f : 0f;
-                var currentAlpha = fadeImage.color.a;
-                float time = 0;
-
-                while (time <= loadDuration) {
-                    time += Time.unscaledDeltaTime;
-                    var alpha = Mathf.Lerp(currentAlpha, targetAlpha, time / loadDuration);
-                    fadeImage.color = new Color(fadeImage.color.r, fadeImage.color.g, fadeImage.color.b, alpha);
-                    await Awaitable.EndOfFrameAsync(token);
-                }
-
-                fadeImage.color = new Color(fadeImage.color.r, fadeImage.color.g, fadeImage.color.b, targetAlpha);
-                onDone?.Invoke();
-            }
-            catch {
-                throw new Exception(TASK_CANCELLATION_MESSAGE);
-            }
+            await token.FadeScreenAsync(image, isFadein ? EFadeType.FadeIn : EFadeType.FadeOut, 
+                duration, Time.unscaledDeltaTime, motion, onDone);
         }
 
         public static async Awaitable<bool> FadeScreenAsync(this CancellationToken token,
-            Image fadeImage,
-            EFadeType fadeType,
-            float loadDuration,
-            float dt,
+            Image image, EFadeType fadeType, float duration, float dt, EMotion motion = EMotion.Linear,
             Action onDone = null) {
             try {
-                if (fadeImage is null) return false;
+                if (image is null) return false;
 
                 var targetAlpha = fadeType switch {
                     EFadeType.FadeOut => 0,
                     EFadeType.FadeIn => 1,
-                    _ => fadeImage.color.a
+                    _ => image.color.a
                 };
 
-                var currentAlpha = fadeImage.color.a;
+                var currentAlpha = image.color.a;
                 float time = 0;
 
-                while (time <= loadDuration) {
+                while (time <= duration) {
                     time += dt;
-                    var alpha = math.lerp(currentAlpha, targetAlpha, time / loadDuration);
-                    fadeImage.color = new Color(fadeImage.color.r, fadeImage.color.g, fadeImage.color.b, alpha);
+                    var alpha = math.lerp(currentAlpha, targetAlpha, motion.Evaluate(time / duration));
+                    image.color = new Color(image.color.r, image.color.g, image.color.b, alpha);
                     await Awaitable.EndOfFrameAsync(token);
                 }
 
-                fadeImage.color = new Color(fadeImage.color.r, fadeImage.color.g, fadeImage.color.b, targetAlpha);
+                image.color = new Color(image.color.r, image.color.g, image.color.b, targetAlpha);
                 onDone?.Invoke();
 
                 return true;
@@ -286,10 +203,7 @@ namespace ChoyUtilities {
         }
 
         public static async Awaitable<bool> FadeScreenAsync(this CancellationToken token,
-            CanvasGroup cg,
-            EFadeType fadeType,
-            float loadDuration,
-            float dt,
+            CanvasGroup cg, EFadeType fadeType, float duration, float dt, EMotion motion = EMotion.Linear,
             Action onDone = null) {
             try {
                 if (cg is null) return false;
@@ -302,9 +216,9 @@ namespace ChoyUtilities {
                 var currentAlpha = cg.alpha;
                 var time = 0f;
 
-                while (time <= loadDuration) {
+                while (time <= duration) {
                     time += dt;
-                    var alpha = math.lerp(currentAlpha, targetAlpha, time / loadDuration);
+                    var alpha = math.lerp(currentAlpha, targetAlpha, motion.Evaluate(time / duration));
                     cg.alpha = alpha;
                     await Awaitable.EndOfFrameAsync(token);
                 }
@@ -323,51 +237,18 @@ namespace ChoyUtilities {
 
         #region Change Color Async
 
-        public static async Awaitable<bool> ChangeColorAsync(this CancellationToken token,
-            Image start,
-            Color target,
-            float loadDuration,
-            Action onDone = null) {
-            try {
-                if (start is null) return false;
-
-                var current = start.color;
-                float time = 0;
-
-                while (time <= loadDuration) {
-                    time += Time.unscaledDeltaTime;
-                    var col = Color.Lerp(current, target, time / loadDuration);
-                    start.color = col;
-                    await Awaitable.EndOfFrameAsync(token);
-                }
-
-                if (token.IsCancellationRequested) return false;
-                start.color = target;
-                onDone?.Invoke();
-
-                return true;
-            }
-            catch {
-                throw new Exception(TASK_CANCELLATION_MESSAGE);
-            }
-        }
-
         public static async Awaitable<Color> ChangeColorAsync(this CancellationToken token,
-            float4 start,
-            float4 target,
-            float loadDuration,
+            float4 start, float4 target, float duration, EMotion motion = EMotion.Linear,
             Action onDone = null) {
             try {
                 float time = 0;
 
-                while (time <= loadDuration) {
+                while (time <= duration) {
                     time += Time.unscaledDeltaTime;
-                    start = math.lerp(start, target, time / loadDuration);
+                    start = math.lerp(start, target, motion.Evaluate(time / duration));
                     await Awaitable.EndOfFrameAsync(token);
                 }
-
                 onDone?.Invoke();
-
                 return new Color(start.x, start.y, start.z, start.w);
             }
             catch {
@@ -376,36 +257,13 @@ namespace ChoyUtilities {
         }
 
         public static async Awaitable<Color> ChangeColorAsync(this CancellationToken token,
-            Color start,
-            Color target,
-            float loadDuration,
+            Color start, Color target, float duration, EMotion motion = EMotion.Linear,
             Action onDone = null) {
-            try {
-                float time = 0;
-                var startColor = new float4(start.r, start.g, start.b, start.a);
-                var targetColor = new float4(target.r, target.g, target.b, target.a);
-                var newColor = new float4();
-
-                while (time <= loadDuration) {
-                    time += Time.unscaledDeltaTime;
-                    newColor = math.lerp(startColor, targetColor, time / loadDuration);
-                    await Awaitable.EndOfFrameAsync(token);
-                }
-
-                onDone?.Invoke();
-
-                return new Color(newColor.x, newColor.y, newColor.z, newColor.w);
-            }
-            catch {
-                throw new Exception(TASK_CANCELLATION_MESSAGE);
-            }
+            return await token.ChangeColorAsync(start, target, duration, Time.unscaledDeltaTime, motion, onDone);
         }
 
         public static async Awaitable<Color> ChangeColorAsync(this CancellationToken token,
-            Color start,
-            Color target,
-            float loadDuration,
-            float dt,
+            Color start, Color target, float duration, float dt, EMotion motion = EMotion.Linear,
             Action onDone = null) {
             try {
                 float time = 0;
@@ -413,12 +271,11 @@ namespace ChoyUtilities {
                 var targetColor = new float4(target.r, target.g, target.b, target.a);
                 var newColor = new float4();
 
-                while (time <= loadDuration) {
+                while (time <= duration) {
                     time += dt;
-                    newColor = math.lerp(startColor, targetColor, time / loadDuration);
+                    newColor = math.lerp(startColor, targetColor, motion.Evaluate(time / duration));
                     await Awaitable.EndOfFrameAsync(token);
                 }
-
                 onDone?.Invoke();
 
                 return new Color(newColor.x, newColor.y, newColor.z, newColor.w);
@@ -433,39 +290,13 @@ namespace ChoyUtilities {
         #region Move Transform Async
 
         public static async Awaitable<bool> MoveAsync(this CancellationToken token,
-            GameObject obj,
-            Transform targetPos,
-            float moveDuration,
+            GameObject obj, Transform targetPos, float duration, EMotion motion = EMotion.Linear,
             Action onDone = null) {
-            try {
-                if (obj is null || targetPos is null) return false;
-
-                var time = 0f;
-                var startPos = obj.transform.position;
-                var endPos = targetPos.position;
-
-                while (time <= moveDuration) {
-                    time += Time.deltaTime;
-                    Vector3 pos = math.lerp(startPos, endPos, time / moveDuration);
-                    obj.transform.position = pos;
-                    await Awaitable.EndOfFrameAsync(token);
-                }
-
-                if (token.IsCancellationRequested) return false;
-                obj.transform.position = endPos;
-                onDone?.Invoke();
-
-                return true;
-            }
-            catch {
-                throw new Exception(TASK_CANCELLATION_MESSAGE);
-            }
+            return await token.MoveAsync(obj, targetPos.position, duration, motion, onDone);
         }
 
         public static async Awaitable<bool> MoveAsync(this CancellationToken token,
-            RectTransform obj,
-            float3 target,
-            float duration,
+            RectTransform obj, float3 target, float duration, EMotion motion = EMotion.Linear,
             Action onDone = null) {
             try {
                 if (obj is null) return false;
@@ -474,7 +305,7 @@ namespace ChoyUtilities {
 
                 while (time <= duration) {
                     time += Time.unscaledDeltaTime;
-                    obj.anchoredPosition3D = math.lerp(start, target, time / duration);
+                    obj.anchoredPosition3D = math.lerp(start, target, motion.Evaluate(time / duration));
                     await Awaitable.NextFrameAsync(token);
                 }
 
@@ -489,9 +320,7 @@ namespace ChoyUtilities {
         }
 
         public static async Awaitable<bool> MoveAsync(this CancellationToken token,
-            GameObject obj,
-            float3 targetPos,
-            float moveDuration,
+            GameObject obj, float3 targetPos, float duration, EMotion motion = EMotion.Linear,
             Action onDone = null) {
             try {
                 if (obj is null) return false;
@@ -499,9 +328,9 @@ namespace ChoyUtilities {
                 float time = 0;
                 var startPos = (float3)obj.transform.position;
 
-                while (time <= moveDuration) {
+                while (time <= duration) {
                     time += Time.unscaledDeltaTime;
-                    var pos = math.lerp(startPos, targetPos, time / moveDuration);
+                    var pos = math.lerp(startPos, targetPos, motion.Evaluate(time / duration));
                     obj.transform.position = pos;
                     await Awaitable.EndOfFrameAsync(token);
                 }